delay from long to int as Calendar.add take int

+++ app/src/main/java/gnn/com/googlealbumdownloadappnougat/photos/SynchronizerDelayedAndroid.java
@@ -8,7 +8,7 @@ import gnn.com.photos.sync.SynchronizerDelayed;
 
 public class SynchronizerDelayedAndroid extends SynchronizerDelayed {
 
-    public SynchronizerDelayedAndroid(long delay, Context activity, File cacheFile, long cacheMaxAge, File processFolder) {
+    public SynchronizerDelayedAndroid(int delay, Context activity, File cacheFile, long cacheMaxAge, File processFolder) {
         super(delay);
         this.synchronizer = new SynchronizerAndroid(activity, cacheFile, cacheMaxAge, processFolder);
     }
